+++ env_compile_all.sh
@@ -1,0 +1,4 @@
+export __COMPILE_MARCHNATIVE=1
+export __O3_OPTIM=1
+export PATH_CKTSO=`pwd`/cktso
+export PATH_NICSLU=`pwd`/nicslu/nicslu202110