+++ lightsim2grid/tests/test_DCPF.py
@@ -45,7 +45,7 @@ class TestDCPF(unittest.TestCase):
         case.name = "case14"
         self._aux_test(case)
 
-    def test_case14_with_phaseshift(self):
+    def test_case14_with_phaseshift_hvside(self):
         # self.skipTest("pandapower might do something weird in DC with phase shift")
         case = pn.case14()        
         hv_bus=0
@@ -65,12 +65,39 @@ class TestDCPF(unittest.TestCase):
                                               vkr_percent=0.0,
                                               shift_degree=-10.0,  # case RTE
                                             #   shift_degree=0.,
-                                              pfe_kw=0.
+                                              pfe_kw=0.,
+                                              tap_side="hv",
                                               )
         # self.tol = 2e-3
         case.name = "case14_2"
         self._aux_test(case)
 
+    def test_case14_with_phaseshift_lvside(self):
+        case = pn.case14()        
+        hv_bus=0
+        lv_bus=2
+        pp.create_transformer_from_parameters(case,
+                                              hv_bus=hv_bus,
+                                              lv_bus=lv_bus,
+                                            #   sn_mva=1184.0,   # case RTE
+                                              sn_mva=9900.0,
+                                              vn_hv_kv=case.bus.iloc[hv_bus]["vn_kv"],
+                                              vn_lv_kv=case.bus.iloc[lv_bus]["vn_kv"],
+                                            #   i0_percent=-0.05152,   # case RTE
+                                              i0_percent=0.0,
+                                            #   vk_percent=0.404445,  # case RTE
+                                              vk_percent=2070.288000,
+                                            #   vkr_percent=0.049728, # case RTE
+                                              vkr_percent=0.0,
+                                              shift_degree=-10.0,  # case RTE
+                                            #   shift_degree=0.,
+                                              pfe_kw=0.,
+                                              tap_side="lv",
+                                              )
+        # self.tol = 2e-3
+        case.name = "case14_3"
+        self._aux_test(case)
+        
     def test_case39(self):
         case = pn.case39()
         case.name = "case39"